pyproject.toml
[tool.poetry]
name = "eknomos-api"
version = "0.1.0"
description = ""
authors = ["Pedro Coelho <pmcoelho@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
django-filter = "^23.3"
django-extensions = "^3.2.3"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.0"
drf-nested-routers = "^0.93.4"
django-lifecycle = "^1.1.2"
psycopg2-binary = "^2.9.9"
django-cors-headers = "^4.3.1"
factory-boy = "^3.3.0"
pypdf2 = "^3.0.1"
pycryptodome = "^3.19.0"
djangorestframework-camel-case = "^1.4.2"
django-colorfield = "^0.10.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
